initial commit all business logic moved to recordOrchestratorService
