+++ service/src/main/java/org/databiosphere/workspacedataservice/controller/RecordController.java
@@ -48,22 +48,12 @@ import java.util.stream.Stream;
 @RestController
 public class RecordController {
 
-	private static final Logger LOGGER = LoggerFactory.getLogger(RecordController.class);
-	private static final int MAX_RECORDS = 1_000;
 	private final RecordDao recordDao;
-	private final DataTypeInferer inferer;
-	private final BatchWriteService batchWriteService;
-	private final RecordService recordService;
+	private final RecordOrchestratorService recordOrchestratorService;
 
-	private final ObjectMapper objectMapper;
-
-	public RecordController(RecordDao recordDao, BatchWriteService batchWriteService, DataTypeInferer inf,
-			ObjectMapper objectMapper, RecordService recordService) {
+	public RecordController(RecordDao recordDao, RecordOrchestratorService recordOrchestratorService) {
 		this.recordDao = recordDao;
-		this.batchWriteService = batchWriteService;
-		this.inferer = inf;
-		this.objectMapper = objectMapper;
-		this.recordService = recordService;
+		this.recordOrchestratorService = recordOrchestratorService;
 	}
 
 	@PatchMapping("/{instanceId}/records/{version}/{recordType}/{recordId}")
@@ -71,22 +61,8 @@ public class RecordController {
 	public ResponseEntity<RecordResponse> updateSingleRecord(@PathVariable("instanceId") UUID instanceId,
 			@PathVariable("version") String version, @PathVariable("recordType") RecordType recordType,
 			@PathVariable("recordId") String recordId, @RequestBody RecordRequest recordRequest) {
-		validateVersion(version);
-		validateInstance(instanceId);
-		checkRecordTypeExists(instanceId, recordType);
-		Record singleRecord = recordDao
-				.getSingleRecord(instanceId, recordType, recordId)
-				.orElseThrow(() -> new MissingObjectException("Record"));
-		RecordAttributes incomingAtts = recordRequest.recordAttributes();
-		RecordAttributes allAttrs = singleRecord.putAllAttributes(incomingAtts).getAttributes();
-		Map<String, DataTypeMapping> typeMapping = inferer.inferTypes(incomingAtts, InBoundDataSource.JSON);
-		Map<String, DataTypeMapping> existingTableSchema = recordDao.getExistingTableSchemaLessPrimaryKey(instanceId, recordType);
-		singleRecord.setAttributes(allAttrs);
-		List<Record> records = Collections.singletonList(singleRecord);
-		Map<String, DataTypeMapping> updatedSchema = batchWriteService.addOrUpdateColumnIfNeeded(instanceId, recordType,
-				typeMapping, existingTableSchema, records);
-		recordService.prepareAndUpsert(instanceId, recordType, records, updatedSchema, recordDao.getPrimaryKeyColumn(recordType, instanceId));
-		RecordResponse response = new RecordResponse(recordId, recordType, singleRecord.getAttributes());
+		RecordResponse response = recordOrchestratorService
+			.updateSingleRecord(instanceId, version, recordType, recordId, recordRequest);
 		return new ResponseEntity<>(response, HttpStatus.OK);
 	}
 
@@ -95,11 +71,7 @@ public class RecordController {
 	public ResponseEntity<RecordResponse> getSingleRecord(@PathVariable("instanceId") UUID instanceId,
 			@PathVariable("version") String version, @PathVariable("recordType") RecordType recordType,
 			@PathVariable("recordId") String recordId) {
-		validateVersion(version);
-		validateInstance(instanceId);
-		checkRecordTypeExists(instanceId, recordType);
-		Record result = recordDao.getSingleRecord(instanceId, recordType, recordId).orElseThrow(() -> new MissingObjectException("Record"));
-		RecordResponse response = new RecordResponse(recordId, recordType, result.getAttributes());
+		RecordResponse response = recordOrchestratorService.getSingleRecord(instanceId, version, recordType, recordId);
 		return new ResponseEntity<>(response, HttpStatus.OK);
 	}
 
@@ -109,15 +81,7 @@ public class RecordController {
 			   @PathVariable("version") String version, @PathVariable("recordType") RecordType recordType,
 			   @RequestParam(name= "primaryKey", required = false) Optional<String> primaryKey,
                @RequestParam("records") MultipartFile records) throws IOException {
-		validateVersion(version);
-		validateInstance(instanceId);
-		if(recordDao.recordTypeExists(instanceId, recordType)){
-			validatePrimaryKey(instanceId, recordType, primaryKey);
-		}
-		int recordsModified;
-		try (InputStreamReader inputStreamReader = new InputStreamReader(records.getInputStream())) {
-			recordsModified = batchWriteService.uploadTsvStream(inputStreamReader, instanceId, recordType, primaryKey);
-		}
+		int recordsModified = recordOrchestratorService.tsvUpload(instanceId, version, recordType, primaryKey, records);
 		return new ResponseEntity<>(new TsvUploadResponse(recordsModified, "Updated " + recordType.toString()),
 				HttpStatus.OK);
 	}
@@ -126,20 +90,8 @@ public class RecordController {
 	// TODO: enable read transaction
 	public ResponseEntity<StreamingResponseBody> streamAllEntities(@PathVariable("instanceId") UUID instanceId,
 			@PathVariable("version") String version, @PathVariable("recordType") RecordType recordType) {
-		validateVersion(version);
-		validateInstance(instanceId);
-		checkRecordTypeExists(instanceId, recordType);
-		List<String> headers = recordDao.getAllAttributeNames(instanceId, recordType);
-
-		StreamingResponseBody responseBody = httpResponseOutputStream -> {
-			try (Stream<Record> allRecords = recordDao.streamAllRecordsForType(instanceId, recordType);
-				 CSVPrinter writer = TsvSupport.getOutputFormat(headers)
-					.print(new OutputStreamWriter(httpResponseOutputStream))) {
-				TsvSupport.RecordEmitter recordEmitter = new TsvSupport.RecordEmitter(writer,
-						headers.subList(1, headers.size()), objectMapper);
-				allRecords.forEach(recordEmitter);
-			}
-		};
+		StreamingResponseBody responseBody =
+			recordOrchestratorService.streamAllEntities(instanceId, version, recordType);
 		return ResponseEntity.status(HttpStatus.OK).contentType(new MediaType("text", "tab-separated-values"))
 				.header(HttpHeaders.CONTENT_DISPOSITION, "attachment;filename=" + recordType.getName() + ".tsv")
 				.body(responseBody);
@@ -151,32 +103,7 @@ public class RecordController {
 			@PathVariable("recordType") RecordType recordType,
 			@PathVariable("version") String version,
 			@RequestBody(required = false) SearchRequest searchRequest) {
-		validateVersion(version);
-		validateInstance(instanceId);
-		checkRecordTypeExists(instanceId, recordType);
-		if (null == searchRequest) {
-			searchRequest = new SearchRequest();
-		}
-		if (searchRequest.getLimit() > MAX_RECORDS || searchRequest.getLimit() < 1 || searchRequest.getOffset() < 0) {
-			throw new ResponseStatusException(HttpStatus.BAD_REQUEST,
-					"Limit must be more than 0 and can't exceed " + MAX_RECORDS + ", and offset must be positive.");
-		}
-		if (searchRequest.getSortAttribute() != null && !recordDao.getExistingTableSchemaLessPrimaryKey(instanceId, recordType)
-				.keySet().contains(searchRequest.getSortAttribute())) {
-			throw new MissingObjectException("Requested sort attribute");
-		}
-		int totalRecords = recordDao.countRecords(instanceId, recordType);
-		if (searchRequest.getOffset() > totalRecords) {
-			return new RecordQueryResponse(searchRequest, Collections.emptyList(), totalRecords);
-		}
-		LOGGER.info("queryForEntities: {}", recordType.getName());
-		List<Record> records = recordDao.queryForRecords(recordType, searchRequest.getLimit(),
-				searchRequest.getOffset(), searchRequest.getSort().name().toLowerCase(),
-				searchRequest.getSortAttribute(), instanceId);
-		List<RecordResponse> recordList = records.stream().map(
-				r -> new RecordResponse(r.getId(), r.getRecordType(), r.getAttributes()))
-				.toList();
-		return new RecordQueryResponse(searchRequest, recordList, totalRecords);
+		return recordOrchestratorService.queryForRecords(instanceId, recordType, version, searchRequest);
 	}
 
 	@PutMapping("/{instanceId}/records/{version}/{recordType}/{recordId}")
@@ -185,34 +112,8 @@ public class RecordController {
 			@PathVariable("version") String version, @PathVariable("recordType") RecordType recordType,
 			@PathVariable("recordId") String recordId, @RequestParam(name= "primaryKey", required = false) Optional<String> primaryKey,
 			 @RequestBody RecordRequest recordRequest) {
-		validateVersion(version);
-		validateInstance(instanceId);
-		RecordAttributes attributesInRequest = recordRequest.recordAttributes();
-		Map<String, DataTypeMapping> requestSchema = inferer.inferTypes(attributesInRequest, InBoundDataSource.JSON);
-		HttpStatus status = HttpStatus.CREATED;
-		if (!recordDao.recordTypeExists(instanceId, recordType)) {
-			RecordResponse response = new RecordResponse(recordId, recordType, recordRequest.recordAttributes());
-			Record newRecord = new Record(recordId, recordType, recordRequest);
-			createRecordTypeAndInsertRecords(instanceId, newRecord, recordType, requestSchema, primaryKey);
-			return new ResponseEntity<>(response, status);
-		} else {
-			validatePrimaryKey(instanceId, recordType, primaryKey);
-			Map<String, DataTypeMapping> existingTableSchema = recordDao.getExistingTableSchemaLessPrimaryKey(instanceId, recordType);
-			// null out any attributes that already exist but aren't in the request
-			existingTableSchema.keySet().forEach(attr -> attributesInRequest.putAttributeIfAbsent(attr, null));
-			if (recordDao.recordExists(instanceId, recordType, recordId)) {
-				status = HttpStatus.OK;
-			}
-			Record newRecord = new Record(recordId, recordType, recordRequest.recordAttributes());
-			List<Record> records = Collections.singletonList(newRecord);
-			batchWriteService.addOrUpdateColumnIfNeeded(instanceId, recordType, requestSchema, existingTableSchema,
-					records);
-			Map<String, DataTypeMapping> combinedSchema = new HashMap<>(existingTableSchema);
-			combinedSchema.putAll(requestSchema);
-			recordService.prepareAndUpsert(instanceId, recordType, records, combinedSchema, primaryKey.orElseGet(() -> recordDao.getPrimaryKeyColumn(recordType, instanceId)));
-			RecordResponse response = new RecordResponse(recordId, recordType, attributesInRequest);
-			return new ResponseEntity<>(response, status);
-		}
+			return recordOrchestratorService.upsertSingleRecord(instanceId, version, recordType, recordId, primaryKey,
+				recordRequest);
 	}
 
 	private void validatePrimaryKey(UUID instanceId, RecordType recordType, Optional<String> primaryKey) {
@@ -224,17 +125,16 @@ public class RecordController {
 	@GetMapping("/instances/{version}")
 	@ReadTransaction
 	public ResponseEntity<List<UUID>> listInstances(@PathVariable("version") String version) {
-		validateVersion(version);
+		RecordOrchestratorService.validateVersion(version);
 		List<UUID> schemaList = recordDao.listInstanceSchemas();
 		return new ResponseEntity<>(schemaList, HttpStatus.OK);
 	}
-
 
 	@PostMapping("/instances/{version}/{instanceId}")
 	@WriteTransaction
 	public ResponseEntity<String> createInstance(@PathVariable("instanceId") UUID instanceId,
 			@PathVariable("version") String version) {
-		validateVersion(version);
+		RecordOrchestratorService.validateVersion(version);
 		if (recordDao.instanceSchemaExists(instanceId)) {
 			throw new ResponseStatusException(HttpStatus.CONFLICT, "This instance already exists");
 		}
@@ -246,8 +146,8 @@ public class RecordController {
 	@WriteTransaction
 	public ResponseEntity<String> deleteInstance(@PathVariable("instanceId") UUID instanceId,
 												 @PathVariable("version") String version) {
-		validateVersion(version);
-		validateInstance(instanceId);
+		RecordOrchestratorService.validateVersion(version);
+		recordOrchestratorService.validateInstance(instanceId);
 		recordDao.dropSchema(instanceId);
 		return new ResponseEntity<>(HttpStatus.OK);
 	}
@@ -257,10 +157,7 @@ public class RecordController {
 	public ResponseEntity<Void> deleteSingleRecord(@PathVariable("instanceId") UUID instanceId,
 			@PathVariable("version") String version, @PathVariable("recordType") RecordType recordType,
 			@PathVariable("recordId") String recordId) {
-		validateVersion(version);
-		validateInstance(instanceId);
-		checkRecordTypeExists(instanceId, recordType);
-		boolean recordFound = recordDao.deleteSingleRecord(instanceId, recordType, recordId);
+		boolean recordFound = recordOrchestratorService.deleteSingleRecord(instanceId, version, recordType, recordId);
 		return recordFound ? new ResponseEntity<>(HttpStatus.NO_CONTENT) : new ResponseEntity<>(HttpStatus.NOT_FOUND);
 	}
 
@@ -268,10 +165,7 @@ public class RecordController {
 	@WriteTransaction
 	public ResponseEntity<Void> deleteRecordType(@PathVariable("instanceId") UUID instanceId,
 			@PathVariable("v") String version, @PathVariable("type") RecordType recordType) {
-		validateVersion(version);
-		validateInstance(instanceId);
-		checkRecordTypeExists(instanceId, recordType);
-		recordDao.deleteRecordType(instanceId, recordType);
+		recordOrchestratorService.deleteRecordType(instanceId, version, recordType);
 		return new ResponseEntity<>(HttpStatus.NO_CONTENT);
 	}
 
@@ -279,10 +173,7 @@ public class RecordController {
 	@ReadTransaction
 	public ResponseEntity<RecordTypeSchema> describeRecordType(@PathVariable("instanceId") UUID instanceId,
 			@PathVariable("v") String version, @PathVariable("type") RecordType recordType) {
-		validateVersion(version);
-		validateInstance(instanceId);
-		checkRecordTypeExists(instanceId, recordType);
-		RecordTypeSchema result = getSchemaDescription(instanceId, recordType);
+		RecordTypeSchema result = recordOrchestratorService.describeRecordType(instanceId, version, recordType);
 		return new ResponseEntity<>(result, HttpStatus.OK);
 	}
 
@@ -290,44 +181,8 @@ public class RecordController {
 	@ReadTransaction
 	public ResponseEntity<List<RecordTypeSchema>> describeAllRecordTypes(@PathVariable("instanceId") UUID instanceId,
 			@PathVariable("v") String version) {
-		validateVersion(version);
-		validateInstance(instanceId);
-		List<RecordType> allRecordTypes = recordDao.getAllRecordTypes(instanceId);
-		List<RecordTypeSchema> result = allRecordTypes.stream()
-				.map(recordType -> getSchemaDescription(instanceId, recordType)).toList();
+		List<RecordTypeSchema> result = recordOrchestratorService.describeAllRecordTypes(instanceId, version);
 		return new ResponseEntity<>(result, HttpStatus.OK);
-	}
-
-	private RecordTypeSchema getSchemaDescription(UUID instanceId, RecordType recordType) {
-		Map<String, DataTypeMapping> schema = recordDao.getExistingTableSchema(instanceId, recordType);
-		List<Relation> relationCols = recordDao.getRelationArrayCols(instanceId, recordType);
-		relationCols.addAll(recordDao.getRelationCols(instanceId, recordType));
-		Map<String, RecordType> relations = relationCols.stream()
-				.collect(Collectors.toMap(Relation::relationColName, Relation::relationRecordType));
-		List<AttributeSchema> attrSchema = schema.entrySet().stream().sorted(Map.Entry.comparingByKey())
-				.map(entry -> createAttributeSchema(entry.getKey(), entry.getValue(), relations.get(entry.getKey())))
-				.toList();
-		int recordCount = recordDao.countRecords(instanceId, recordType);
-		return new RecordTypeSchema(recordType, attrSchema, recordCount, recordDao.getPrimaryKeyColumn(recordType, instanceId));
-	}
-
-	private AttributeSchema createAttributeSchema(String name, DataTypeMapping datatype, RecordType relation) {
-		if (relation == null) {
-			return new AttributeSchema(name, datatype.toString(), null);
-		}
-		return new AttributeSchema(name, "STRING".equals(datatype.toString()) ? "RELATION" : "ARRAY_OF_RELATION", relation);
-	}
-
-	private static void validateVersion(String version) {
-		if (null == version || !version.equals("v0.2")) {
-			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Invalid API version specified");
-		}
-	}
-
-	private void checkRecordTypeExists(UUID instanceId, RecordType recordType) {
-		if (!recordDao.recordTypeExists(instanceId, recordType)) {
-			throw new MissingObjectException("Record type");
-		}
 	}
 
 	@PostMapping("/{instanceid}/batch/{v}/{type}")
@@ -335,26 +190,7 @@ public class RecordController {
 	public ResponseEntity<BatchResponse> streamingWrite(@PathVariable("instanceid") UUID instanceId,
 			@PathVariable("v") String version, @PathVariable("type") RecordType recordType,
 			@RequestParam(name= "primaryKey", required = false) Optional<String> primaryKey, InputStream is) {
-		validateVersion(version);
-		validateInstance(instanceId);
-		if(recordDao.recordTypeExists(instanceId, recordType)){
-			validatePrimaryKey(instanceId, recordType, primaryKey);
-		}
-		int recordsModified = batchWriteService.consumeWriteStream(is, instanceId, recordType, primaryKey);
+		int recordsModified = recordOrchestratorService.streamingWrite(instanceId, version, recordType, primaryKey, is);
 		return new ResponseEntity<>(new BatchResponse(recordsModified, "Huzzah"), HttpStatus.OK);
 	}
-
-	private void validateInstance(UUID instanceId) {
-		if (!recordDao.instanceSchemaExists(instanceId)) {
-			throw new MissingObjectException("Instance");
-		}
-	}
-
-	private void createRecordTypeAndInsertRecords(UUID instanceId, Record newRecord, RecordType recordType,
-		Map<String, DataTypeMapping> requestSchema, Optional<String> primaryKey) {
-		List<Record> records = Collections.singletonList(newRecord);
-		recordDao.createRecordType(instanceId, requestSchema, recordType, inferer.findRelations(records, requestSchema), primaryKey.orElse(ReservedNames.RECORD_ID));
-		recordService.prepareAndUpsert(instanceId, recordType, records, requestSchema, primaryKey.orElse(ReservedNames.RECORD_ID));
-	}
-
 }
